Relocate estimation of mantle thickness in ParameterIO.hpp
Different radii for the core mantle boundary (rCMB) and Earth's surface (rSurface) can be defined within the parameters.prm file.
If radial profiles are utilized it can occur that the calculated mantle thickness derived from rCMB and rSurface is
overwritten by a default value within ParametersIO.hpp and hence incorrectly estimated non-dimensional values for rMax and rMin might be derived.
In order to avoid this, the mantle thickness estimation is relocated.
With this adjustment it is ensured that the correct mantle thickness will be utilized for all further estimations/non-dimensionalisations.