Add PETSc MinRes as an option for the coarse grid solver
Users can now define if the HyTeG MinRes coarse grid solver will be used or if a PETSc MinRes coarse grid solver should be utilized in TerraNeo simulation runs.
By defining the corresponding flag "PETScFlag" within the parameters.prm file the coarse grid solver can be set.
If a PETSc MinRes solver is chosen it must be ensured that TerraNeo is build with PETSc being enabled.
The amount of PETSc coarse grid iterations is derived from the solver coarse grid iteration parameter in the parameters file.