Minor behavioral changes in MicroMesh.
Now by default, if a GeometryMap is present when the MicroMesh is initialized, that GeometryMap is applied and interpolated into the MicroMesh.

Functions for interpolation of the refined mesh now force the user to decide explicitly whether a present GeometryMap shall be ignored or not.

Updated show_mesh app since now these function calls can be removed.