Add option to utilize reference temperature profiles in TerraNeo
A reference temperature profile can be provided within the parameters.prm file.
If a profile is entered, the application will use this profile as the reference temperature profile throughout the mantle.
The temperature deviations will henceforth be estimated with respect to this profile.
The profile has to be in .json format with one entry holding the key "Radius (m)" and the corresponding
data values in descending order starting from the surface.
The second entry must be the temperature.