Add an MPITagProvider class
The MPITagProvider class is used by BufferedCommunicator to obtain a value
for an MPI tag to pass to walberla::OpenMPBufferSystem. It uses the same
simple approach for generating unique tags, which is to increment the counter
after each query. However, it also verifies that the tag stays within the
available range provided by the MPI library in use. See issue #282 for why
this might be important.