Move some implementation from DGFunction "down" to VolumeDoFFunction
The VolumeDoFFunction conceptually has no idea what its degrees of freedom
represent, only the FE functions that build on it do. However, for the
methods getMaxDoFValue(), getMinDoFValue() and getMaxDoFMagnitude() such
knowledge is not necessary.

This commit adds an implementation of the three methods to VolumeDoFFunction
and makes their DGFunction counterparts delegate the work to these.