Implements micromesh::microFaceCenterPosition()
This free-function determines the "center" of a micro-face. Here we
define as "center" the location of the barycenter of the reference
simplex after it was mapped first to the computational and then to the
physical domain, where the latter mapping can either be affine (degree 1),
quadratic (degree 2) or a blending map.