Implements MicroMeshPositionTest
The new test extends testing of the MicroMesh functionality. In details
we check that the following free-functions from the micromesh namespace
work correctly:

- micromesh::microVertexPosition()
- micromesh::microEdgeCenterPosition() <-- version using EdgeDoFIndices
- micromesh::microFaceCenterPosition()

The testing is only done in 2D so far, but for all currently supported cases,
i.e.

a) no blending or micromesh
b) blending
c) use of degree 1 MicroMesh object
d) use of degree 2 MicroMesh object