Implement radial profiles for density and update non-dimensional parameters
Density radial profiles can now be loaded and used in TerraNeo.
A user specified density profile in .JSON format can be loaded and
used instead of the exponentional function to derive a background radial density profile.
The JSON file must contain one entry with the key "Radius (m)" and values in descending order
while the second entry must contain the corresponding values (e.g. density in kg/m^3).
Users must ensure that the min and max radius of the loaded radial profile are consistent with
the domain min and max radius defined in the parameters file.
Also, a more generalized method to load radial profiles is implemented.

Additonally, the non-dimensional parameters Ra (Rayleigh-number), Di (Dissipation-number)
and Pe (Peclet-number) are no longer directly updated if radial profiles for the thermal
expansivity, specific heat capacity and grueneisen parameter are provided but are scaled
by the respective non-dimensional forms for the thermal expansivity, specific heat capacity
and grueneisen parameter according to the non-dimensional forms within the Stokes
and energy equation described in Schubert and Turcotte 2001 (Chapter 6.10).