Implements DG1Function::interpolate() for constant values
Implementation was easy, we just delegate to DGFunction::interpolate(),
where interpolation of constants is already implemented.

Funny aspect: DGPoisson2DConvergenceTest failed before. It uses a DG1Function
and because the CG solver tries to call interpolate for a constant on this,
we got a "not implemented" abort. DGPoisson3DConvergenceTest, however, worked.
It also uses a DG1 FE function, but not as object of DG1Function class, but
"manually" through a DGFunction object.