Update c'tors of [DG|VolumeDoF]Function w.r.t. additionalHaloDepth
Before this commit, construction of a VolumeDoFFunction failed, if the
underlying PrimitiveStorage did not have an additionalHaloDepth > 0. The
direct reason for this is the call of the class method
communicateNumScalarsPerPrimitive(). The commit avoids calling the latter,
for additionalHaloDepth = 0. The c'tor of DGFunction will instead log a
warning for the user.

Running the P2PlusBubbleFunctionTest works fine without additional halo
depth. This is expected, as the bubble dof (represented by a
VolumeDoFFunction) does not couple to dofs on other elements.

See also issue #289.