Implement incompressibility check in TerraNeo
When calculating the non-dimensional density via densityFunction(x)
for incompressible simulation runs, TerraNeo was still utilizing
either a density profile or an adiabatic compression function, depending
on the input parameters.
This is not correct. In case of incompressible runs non-dimensional
density must be evaluated with 1.
An additional check was implemented that ensures density is set to 1
if an incompressible run is started in TerraNeo.