Update safety checks in DGFunction::interpolate()
Interpolation for a constant value, so far, was restricted in 3D to
the case of DG0 (a.k.a. P0), while there was no check for 2D.

This commit instead tests, whether basis_ is a nodal basis, i.e. of
type DGBasisLinearLagrange_Example. If this is true, we set all DoFs
to the prescribed constant value. If not, we abort.