Implement DG1Function::interpolate() for expressions
DG1Function uses a standard nodal basis for the linear polynomial on
each element. Based on this we can implement interpolation in a straight-
forward manner.

Note that the evaluation of the expression already uses the micromesh
functionality. A step in direction of resolving issue #293.

The commit extends DGInterpolateEvaluateTest for some basic testing.