Introduces a free-function hyteg::buildinfo::fpType()
The new function will hand back a string representing the type that is
behind real_t. This info is then used in printBuildInfo() and
printSoftwareMetaData(). The PrintInfoTest is extended accordingly.

The commit additionally updates the reference files for the
AdiosWriterVerifyTest, fixes old parameter names in the usage info and
extends a docstring in the python file. It also adds a shell-script to
help re-create the reference files in the future.