Updates P2PlusBubbleFunctionTest to use operators from hyteg_operators
The commit changes P2PlusBubbleFunctionTest to use the generated operators
from hyteg_operators instead of those from the folder
hyteg/experimental/P2PlusBubbleOperators. It removes the latter operators.
Only exception for the moment is the mass operator for the AffineMap2D,
which we keep for the moment, but place in tests/hyteg/P2.

As a consequence, we can run the test then also for single precision,
excluding the AffineMap2D case then.