Added MicroElementDoFTypification.hpp and a related test MicroElementDoFTypificationTest.cpp.
MicroElementDoFTypification.hpp contains functions that allow the user to determine the DoFType of degrees of freedom of a given micro element (P2 or P1). This was developed to allow the user to construct local linear systems potentially spanning multiple elements. Could also be potentially used within form evaluation. The implementation only uses indexing information and switch cases, allowing for fast and efficient typification.