Moves check for additional halo depth to DGOperator class
The commit moves the check for additional halo depth to DGOperator class.
Before it was in the DGFunction constructor. However, whether we need an
additional halo depth > 0 depends on the form. The DGOperator IMHO is a
better place to check this, as it can query the form for having other than
volume integrals and also has access to the PrimitiveStorage object.