Added createInnerBC to BoundaryConditions.cpp/.hpp.
Expanded upon possible user attribute types in the Adios Importer/Exporter. Unified with the independently developed changes in the MC App. Updated the respective test to also test the additional features.

Added functionality to BufferedCommunication.cpp/.hpp and MPITagProvider.cpp/.hpp to automatically returned unused MPITags back to a reusable pool.

Added using valueType = ValueType and setToZero to serveral function types that were missing the respective parts.

Changed how copyBoundaryConditionsFromFunction works for GenericFunction.

Set getMinLevel and getMaxLevel to const and fixed copyBoundaryConditionFromFunction for CSFVectorFunction.

Added pressureMeanProjection.hpp that allows for a function type based projectMean call.

Added the FunctionHistory class. A class that can hold an arbitrary amount of past states (with time steps and potential custom data attached) and allows for automatic extrapolation and calculation of time derivative approximations.

Added an additional closest point distance to P1Function and P2Function evaluate ( and the respective helper functions ). This allows the user to set a distance tolerance to find a point pairing element. This allows evaluate to accept point pairings as long as the point to element distance is small enough. Set to zero (default) to get the same behaviour as before.

Added a bestGuess option to P1Function and P2Function evaluate. If set to true (false by default) evaluate returns the best guess, i.e., an evaluation at the computational domain point with the smallest point to element distance, that also fulfills the point pairing criterion.

Added a temporary function manager in TempFunctionManager.hpp. This allows the user to draw an arbitrary function type temporary function from a pool that gets returned to the pool (and potentially reused) if the function runs out of scope.

Added a refinedCoarseMesh2D function to MeshGenRefineCoarseGrid.hpp/.cpp.

Added an "Substitution failure is not an error" toolbox that allows the user to check (at compile time) if an operator/class defines certain methods.

Added a NoOperator dummy class, used in the MCApp.

Added a FunctionHistory based TimeDiscretisation class and a derived variable time step size BDF-K and CrankNicolson scheme. This allows for an easy implementation of time discretisations without having to program them by hand each time.

Added a GEMV manager. The included functions allow the user to apply, toMatrix, smooth_jac and computeInverseDiagonalOperatorValues in a scaled fashion for a given operator of choice. The functions automatically check for efficient applyScaled, toMatrixScaled, smooth_jac, gemv, ... functions defined for the operator via SFINAE and if none are available it calculates the result via temporary functions / matrix calculations (less efficient).

Added a ScaledOperator wrapper using the above mentioned GEMV tools and SFINAE.

Added serveral missing communication routines for P1 Functions to VertexDoFAdditivePackInfo.hpp. Note that the way to handle communication in these cases is ambigous and could also be handled differently ( i.e. which DoFs communicate in what way ).

Added some MPI_Barriers to the roundRobin method in distributedBalancer.cpp. This has proven to be potentially necessary on large clusters.

Added a pre solve call of stopIterationCallback for solverLoop.hpp.

Added an IdentityOperator.

Added ApplyInverseDiagonalWrapper.hpp, wrapping multiplication with the inverse diagonal as an operator with an apply method.

Added FunctionMultiplicationPreconditioner, wrapping multiplication with a given function as a solver / preconditioner.

Fixed some solvers missing hyteg::FreeSlipBoundary in their flag_.

Added the ConvectionToolbox. A toolbox / wrapper for easy creation of mantle convection storages.

Added closestPointTriangle3D and closestPointTetrahedron3D functions and a associated test.

Added ElementwiseAndConstantOperatorScalingTest, testing the functionality of GEMV.hpp and ScaledOperator.hpp.