Refactors primitive getter methods of SetupPrimitiveStorage class
The commit changes the return type of getVertex() and the like to be
a shared_ptr instead of a raw pointer. Yes, this involves copying the
shared pointer stored in the VertexMap, which has a certain overhead.
However, as we also generate copies of the map itself, I do not consider
this a big problem, especially not in the case of SetupPrimitiveStorage.