Adds basic functionality to create and work on a subset of a PrimitiveStorage
The commit adds functionality to extract a subset from an existing
SetupPrimitiveStorage. This can in turn be used to generate a new
PrimitiveStorage object representing a "submesh".

The primitives of the "SubStorage" will reside on the same MPI processes
as their counterparts from the "SuperStorage".

The commit also provides a assignAcrossStorages() function that allows to
assign a weighted linear combination of source functions to a destination
function. The functions can live on different PrimitiveStorages. The only
requirement is that the "targetStorage" on which the assignment is intended
to happen forms a subset of the PrimitiveStorages of all functions involved.