Allows setting the format of ADIOS2 checkpoints via CMake
The commit renames the ADIOS2_BP_FORMAT macro to ADIOS2_CHECKPOINT_FORMAT
and makes it an advanced option whicht can be set with CMake. We also update
the default from BP4 to BP5.

The AdiosCheckpointImporter and also the PrimitiveStorage now use as engine
type the generic "BPFile", this is ADIOS2's default anyway.