Refactor adiosCheckpointHelpers_v03::deserializePrimitiveMap()
The previous implementation was using an std::streambuffer, which was
prone to failure. This implementation now directly uses a new
PrimitiveID::fromBuffer() to which we pass an std::vector and a position
inside of the latter.

Additionally we let the function check for the success of inserting the
new pair into the map.